refactor #20: DTO 수정
