--- src/main/java/org/prgms/locomocoserver/categories/dto/response/AllCategoriesWithTagsResponseDto.java
@@ -1,8 +1,0 @@
-package org.prgms.locomocoserver.categories.dto.response;
-
-import java.util.List;
-import org.prgms.locomocoserver.categories.dto.CategoriesWithTagsDto;
-
-public record AllCategoriesWithTagsResponseDto(List<CategoriesWithTagsDto> data) {
-
-}
