+++ src/main/java/org/prgms/locomocoserver/global/common/dto/Results.java
@@ -1,0 +1,7 @@
+package org.prgms.locomocoserver.global.common.dto;
+
+import java.util.List;
+
+public record Results<T>(List<T> data) {
+
+}
